move all profile-related code to one profile.rs
