+++ src/core/mod.rs
@@ -1,5 +1,3 @@
 pub mod auth;
 pub mod error;
-pub mod inner;
 pub mod models;
-pub mod rpc;
