+++ src/services/mod.rs
@@ -1,5 +1,4 @@
 pub mod home;
-pub mod profile;
 pub mod relationship;
 pub mod search;
 pub mod tweet;
